Refactor gestalt CLI agents module and deduplicate turn handling.
Split the monolithic agents command into focused modules, share turn-event
parsers and a turn driver between stdout and TUI paths, and centralize query
helpers plus workflow dispatch to reduce duplication.
