--- gestalt/cli/src/commands/agents.rs
@@ -1,2811 +1,0 @@
-use anyhow::{Context, Result, bail};
-use colored::Colorize;
-use serde::{Deserialize, Deserializer, Serialize, de::DeserializeOwned};
-use serde_json::{Map, Value, json};
-use std::io::{self, BufRead, BufReader, IsTerminal, Write};
-use std::path::{Path, PathBuf};
-use std::process::{Command as ProcessCommand, ExitStatus, Stdio};
-use std::sync::{
-    Arc,
-    atomic::{AtomicBool, AtomicU64, Ordering},
-};
-use std::thread;
-use std::time::Duration;
-use time::{OffsetDateTime, format_description::well_known::Rfc3339};
-
-use crate::api::ApiClient;
-use crate::cli::{
-    AgentArgs, AgentResumeArgs, AgentSessionCreateArgs, AgentSessionUpdateArgs, AgentToolArg,
-    AgentTurnCreateArgs, AgentTurnEventListArgs, AgentTurnEventStreamArgs,
-};
-use crate::interactive::{
-    InputPrompt, InteractiveLineReader, PromptLine, prompt_confirm, prompt_input,
-};
-use crate::output::{self, Format};
-use crate::params;
-
-mod display_markdown;
-mod tui;
-
-const SESSIONS_PATH: &str = "/api/v1/agent/sessions";
-const PROVIDERS_PATH: &str = "/api/v1/agent/providers";
-const HARNESSES_RESOLVE_PATH: &str = "/api/v1/agent/harnesses/resolve";
-const TURNS_PATH: &str = "/api/v1/agent/turns";
-const DEFAULT_SESSION_LIST_LIMIT: usize = 50;
-const DEFAULT_EVENT_PAGE_SIZE: u32 = 100;
-const EVENT_POLL_INTERVAL: Duration = Duration::from_millis(250);
-const EVENT_STREAM_UNTIL_BLOCKED_OR_TERMINAL: &str = "blocked_or_terminal";
-const INTERRUPT_CANCEL_REASON: &str = "operator interrupted";
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentHarnessResolveRequest<'a> {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    provider: Option<&'a str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    harness: Option<&'a str>,
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentHarnessPlan {
-    provider: String,
-    #[serde(default)]
-    harness: String,
-    command: String,
-    #[serde(default)]
-    args: Vec<String>,
-    #[serde(default)]
-    env: Map<String, Value>,
-    #[serde(default)]
-    working_directory: String,
-    #[serde(default)]
-    required_commands: Vec<String>,
-    install: Option<AgentHarnessInstallPlan>,
-}
-
-#[derive(Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-struct AgentHarnessInstallPlan {
-    #[serde(default)]
-    instructions: String,
-    #[serde(default)]
-    commands: Vec<AgentHarnessInstallCommand>,
-}
-
-#[derive(Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-struct AgentHarnessInstallCommand {
-    #[serde(default)]
-    description: String,
-    #[serde(default)]
-    command: String,
-    #[serde(default)]
-    args: Vec<String>,
-    #[serde(default)]
-    shell: String,
-    #[serde(default)]
-    env: Map<String, Value>,
-}
-
-pub fn launch_local(
-    client: &ApiClient,
-    provider: Option<&str>,
-    harness: Option<&str>,
-) -> Result<()> {
-    let plan = resolve_agent_harness(client, provider, harness)?;
-    let prepared = prepare_agent_harness(&plan, HarnessInstallPolicy::Prompt)?;
-    launch_agent_harness(&plan, &prepared)
-}
-
-pub fn doctor_local(
-    client: &ApiClient,
-    provider: Option<&str>,
-    harness: Option<&str>,
-) -> Result<()> {
-    let plan = resolve_agent_harness(client, provider, harness)?;
-    let _ = prepare_agent_harness(&plan, HarnessInstallPolicy::CheckOnly)?;
-    let harness_label = if plan.harness.trim().is_empty() {
-        plan.provider.clone()
-    } else {
-        format!("{}/{}", plan.provider, plan.harness)
-    };
-    println!("agent harness {:?} ok", harness_label);
-    Ok(())
-}
-
-fn resolve_agent_harness(
-    client: &ApiClient,
-    provider: Option<&str>,
-    harness: Option<&str>,
-) -> Result<AgentHarnessPlan> {
-    let body = AgentHarnessResolveRequest { provider, harness };
-    let resp = client
-        .post(HARNESSES_RESOLVE_PATH, &body)
-        .context("failed to resolve agent harness")?;
-    serde_json::from_value(resp).context("invalid agent harness response")
-}
-
-struct PreparedAgentHarness {
-    command: PathBuf,
-    env: Vec<String>,
-    working_directory: Option<PathBuf>,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum HarnessInstallPolicy {
-    CheckOnly,
-    Prompt,
-}
-
-struct MissingHarnessCommand {
-    command: String,
-    role: MissingHarnessCommandRole,
-    detail: String,
-}
-
-#[derive(Clone, Copy)]
-enum MissingHarnessCommandRole {
-    Required,
-    HarnessCommand,
-}
-
-fn prepare_agent_harness(
-    plan: &AgentHarnessPlan,
-    install_policy: HarnessInstallPolicy,
-) -> Result<PreparedAgentHarness> {
-    let mut attempted_install = false;
-    loop {
-        let prepared = prepare_agent_harness_once(plan, install_policy, attempted_install)?;
-        match prepared {
-            PrepareAgentHarnessResult::Ready(prepared) => return Ok(prepared),
-            PrepareAgentHarnessResult::Installed => {
-                attempted_install = true;
-            }
-        }
-    }
-}
-
-enum PrepareAgentHarnessResult {
-    Ready(PreparedAgentHarness),
-    Installed,
-}
-
-fn prepare_agent_harness_once(
-    plan: &AgentHarnessPlan,
-    install_policy: HarnessInstallPolicy,
-    attempted_install: bool,
-) -> Result<PrepareAgentHarnessResult> {
-    let env = effective_harness_env(&plan.env)?;
-    let working_directory = if plan.working_directory.trim().is_empty() {
-        None
-    } else {
-        let path = PathBuf::from(plan.working_directory.trim());
-        let metadata = std::fs::metadata(&path).with_context(|| {
-            format!(
-                "agent harness working directory {:?} is not accessible",
-                path.display().to_string()
-            )
-        })?;
-        if !metadata.is_dir() {
-            bail!(
-                "agent harness working directory {:?} is not a directory",
-                path.display().to_string()
-            );
-        }
-        Some(path)
-    };
-
-    let missing = missing_harness_commands(plan, working_directory.as_deref(), &env);
-    if !missing.is_empty() {
-        if install_policy == HarnessInstallPolicy::Prompt
-            && !attempted_install
-            && maybe_run_agent_harness_install(plan, &missing, working_directory.as_deref(), &env)?
-        {
-            return Ok(PrepareAgentHarnessResult::Installed);
-        }
-        bail!("{}", missing_harness_commands_message(plan, &missing));
-    }
-    let command = find_command(plan.command.trim(), working_directory.as_deref(), &env)
-        .with_context(|| format!("agent harness command {:?} is not available", plan.command))?;
-    Ok(PrepareAgentHarnessResult::Ready(PreparedAgentHarness {
-        command,
-        env: env_list(env),
-        working_directory,
-    }))
-}
-
-fn missing_harness_commands(
-    plan: &AgentHarnessPlan,
-    working_directory: Option<&Path>,
-    env: &std::collections::BTreeMap<String, String>,
-) -> Vec<MissingHarnessCommand> {
-    let mut missing = Vec::new();
-    for required in &plan.required_commands {
-        let required = required.trim();
-        if required.is_empty() {
-            continue;
-        }
-        if let Err(err) = find_command(required, working_directory, env) {
-            missing.push(MissingHarnessCommand {
-                command: required.to_string(),
-                role: MissingHarnessCommandRole::Required,
-                detail: err.to_string(),
-            });
-        }
-    }
-    let command = plan.command.trim();
-    if missing.iter().any(|missing| missing.command == command) {
-        return missing;
-    }
-    if let Err(err) = find_command(command, working_directory, env) {
-        missing.push(MissingHarnessCommand {
-            command: command.to_string(),
-            role: MissingHarnessCommandRole::HarnessCommand,
-            detail: err.to_string(),
-        });
-    }
-    missing
-}
-
-fn maybe_run_agent_harness_install(
-    plan: &AgentHarnessPlan,
-    missing: &[MissingHarnessCommand],
-    working_directory: Option<&Path>,
-    env: &std::collections::BTreeMap<String, String>,
-) -> Result<bool> {
-    let Some(install) = plan.install.as_ref() else {
-        return Ok(false);
-    };
-    if install.commands.is_empty() {
-        return Ok(false);
-    }
-    if !io::stdin().is_terminal() || !io::stderr().is_terminal() {
-        return Ok(false);
-    }
-
-    eprintln!("{}", missing_harness_commands_summary(missing));
-    if let Some(guidance) = install_guidance(plan) {
-        eprintln!();
-        eprintln!("{guidance}");
-    }
-    let question = if install.commands.len() == 1 {
-        "Run install command now?"
-    } else {
-        "Run install commands now?"
-    };
-    if !prompt_confirm(question, false)? {
-        return Ok(false);
-    }
-
-    for command in &install.commands {
-        run_agent_harness_install_command(command, working_directory, env)?;
-    }
-    Ok(true)
-}
-
-fn run_agent_harness_install_command(
-    install: &AgentHarnessInstallCommand,
-    working_directory: Option<&Path>,
-    env: &std::collections::BTreeMap<String, String>,
-) -> Result<()> {
-    let mut cmd = if !install.shell.trim().is_empty() {
-        shell_command(install.shell.trim())
-    } else {
-        let command = install.command.trim();
-        if command.is_empty() {
-            bail!("agent harness install command is required");
-        }
-        let mut cmd = ProcessCommand::new(command);
-        cmd.args(&install.args);
-        cmd
-    };
-    if let Some(working_directory) = working_directory {
-        cmd.current_dir(working_directory);
-    }
-    cmd.env_clear();
-    cmd.envs(
-        env.iter()
-            .map(|(key, value)| (key.as_str(), value.as_str())),
-    );
-    for (key, value) in &install.env {
-        let Some(value) = value.as_str() else {
-            bail!("agent harness install env {:?} must be a string", key);
-        };
-        cmd.env(key, value);
-    }
-    let display = install_command_display(install);
-    let status = cmd
-        .stdin(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .status()
-        .with_context(|| format!("failed to start agent harness install command: {display}"))?;
-    if !status.success() {
-        bail!(
-            "agent harness install command failed with status {}: {}",
-            helper_exit_code(status),
-            display
-        );
-    }
-    Ok(())
-}
-
-fn shell_command(script: &str) -> ProcessCommand {
-    #[cfg(windows)]
-    {
-        let mut cmd = ProcessCommand::new("cmd");
-        cmd.args(["/C", script]);
-        cmd
-    }
-    #[cfg(not(windows))]
-    {
-        let mut cmd = ProcessCommand::new("sh");
-        cmd.args(["-c", script]);
-        cmd
-    }
-}
-
-fn missing_harness_commands_message(
-    plan: &AgentHarnessPlan,
-    missing: &[MissingHarnessCommand],
-) -> String {
-    let mut message = missing_harness_commands_summary(missing);
-    if let Some(guidance) = install_guidance(plan) {
-        message.push_str("\n\n");
-        message.push_str(&guidance);
-    }
-    message
-}
-
-fn missing_harness_commands_summary(missing: &[MissingHarnessCommand]) -> String {
-    if missing.len() == 1 {
-        let missing = &missing[0];
-        return match missing.role {
-            MissingHarnessCommandRole::Required => format!(
-                "required command {:?} is not available: {}",
-                missing.command, missing.detail
-            ),
-            MissingHarnessCommandRole::HarnessCommand => format!(
-                "agent harness command {:?} is not available: {}",
-                missing.command, missing.detail
-            ),
-        };
-    }
-    let mut message = "agent harness commands are not available:".to_string();
-    for missing in missing {
-        message.push_str(&format!(
-            "\n  - {}: {}",
-            install_command_quote(&missing.command),
-            missing.detail
-        ));
-    }
-    message
-}
-
-fn install_guidance(plan: &AgentHarnessPlan) -> Option<String> {
-    let install = plan.install.as_ref()?;
-    let instructions = install.instructions.trim();
-    if instructions.is_empty() && install.commands.is_empty() {
-        return None;
-    }
-
-    let mut lines = vec![format!(
-        "Install instructions for agent harness {:?}:",
-        harness_label(plan)
-    )];
-    if !instructions.is_empty() {
-        lines.push(instructions.to_string());
-    }
-    for command in &install.commands {
-        if !command.description.trim().is_empty() {
-            lines.push(format!("{}:", command.description.trim()));
-        }
-        lines.push(format!("  $ {}", install_command_display(command)));
-    }
-    Some(lines.join("\n"))
-}
-
-fn harness_label(plan: &AgentHarnessPlan) -> String {
-    if plan.harness.trim().is_empty() {
-        plan.provider.clone()
-    } else {
-        format!("{}/{}", plan.provider, plan.harness)
-    }
-}
-
-fn install_command_display(command: &AgentHarnessInstallCommand) -> String {
-    if !command.shell.trim().is_empty() {
-        return command.shell.trim().to_string();
-    }
-    let mut parts = vec![install_command_quote(command.command.trim())];
-    parts.extend(command.args.iter().map(|arg| install_command_quote(arg)));
-    parts.join(" ")
-}
-
-fn install_command_quote(value: &str) -> String {
-    if value.is_empty() {
-        return "''".to_string();
-    }
-    if value
-        .bytes()
-        .all(|b| b.is_ascii_alphanumeric() || b"@%_+=:,./-".contains(&b))
-    {
-        return value.to_string();
-    }
-    format!("'{}'", value.replace('\'', "'\\''"))
-}
-
-fn launch_agent_harness(plan: &AgentHarnessPlan, prepared: &PreparedAgentHarness) -> Result<()> {
-    let mut cmd = ProcessCommand::new(&prepared.command);
-    cmd.args(&plan.args);
-    if let Some(working_directory) = &prepared.working_directory {
-        cmd.current_dir(working_directory);
-    }
-    cmd.env_clear();
-    cmd.envs(
-        prepared
-            .env
-            .iter()
-            .filter_map(|entry| entry.split_once('=')),
-    );
-    let status = cmd
-        .stdin(Stdio::inherit())
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .status()
-        .with_context(|| format!("failed to start agent harness {:?}", plan.provider))?;
-    if status.success() {
-        return Ok(());
-    }
-    std::process::exit(helper_exit_code(status));
-}
-
-fn helper_exit_code(status: ExitStatus) -> i32 {
-    if let Some(code) = status.code() {
-        return code;
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::process::ExitStatusExt;
-        if let Some(signal) = status.signal() {
-            return 128 + signal;
-        }
-    }
-    1
-}
-
-fn effective_harness_env(
-    overlay: &Map<String, Value>,
-) -> Result<std::collections::BTreeMap<String, String>> {
-    let mut env: std::collections::BTreeMap<String, String> = std::env::vars().collect();
-    for (key, value) in overlay {
-        let Some(value) = value.as_str() else {
-            bail!("agent harness env {:?} must be a string", key);
-        };
-        env.insert(key.clone(), value.to_string());
-    }
-    Ok(env)
-}
-
-fn env_list(env: std::collections::BTreeMap<String, String>) -> Vec<String> {
-    env.into_iter()
-        .map(|(key, value)| format!("{key}={value}"))
-        .collect()
-}
-
-fn find_command(
-    command: &str,
-    working_directory: Option<&Path>,
-    env: &std::collections::BTreeMap<String, String>,
-) -> Result<PathBuf> {
-    if command.trim().is_empty() {
-        bail!("command is required");
-    }
-    let candidate = PathBuf::from(command);
-    if has_path_separator(command) {
-        let resolved = if candidate.is_absolute() {
-            candidate
-        } else if let Some(working_directory) = working_directory {
-            working_directory.join(candidate)
-        } else {
-            std::env::current_dir()
-                .context("failed to resolve current directory")?
-                .join(candidate)
-        };
-        if is_executable_file(&resolved) {
-            return Ok(resolved);
-        }
-        bail!("{command} not found");
-    }
-
-    let path_value = env.get("PATH").map(String::as_str).unwrap_or("");
-    for dir in std::env::split_paths(path_value) {
-        let dir = if dir.as_os_str().is_empty() {
-            PathBuf::from(".")
-        } else {
-            dir
-        };
-        let dir = if !dir.is_absolute() {
-            working_directory
-                .map(|working_directory| working_directory.join(&dir))
-                .unwrap_or(dir)
-        } else {
-            dir
-        };
-        let resolved = dir.join(command);
-        if is_executable_file(&resolved) {
-            return Ok(resolved);
-        }
-        #[cfg(windows)]
-        {
-            let resolved = dir.join(format!("{command}.exe"));
-            if is_executable_file(&resolved) {
-                return Ok(resolved);
-            }
-        }
-    }
-    bail!("{command} not found in PATH");
-}
-
-fn has_path_separator(command: &str) -> bool {
-    command.contains('/')
-        || if cfg!(windows) {
-            command.contains('\\')
-        } else {
-            false
-        }
-}
-
-fn is_executable_file(path: &Path) -> bool {
-    let Ok(metadata) = std::fs::metadata(path) else {
-        return false;
-    };
-    if !metadata.is_file() {
-        return false;
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        metadata.permissions().mode() & 0o111 != 0
-    }
-    #[cfg(not(unix))]
-    {
-        true
-    }
-}
-
-pub fn create_session(
-    client: &ApiClient,
-    args: &AgentSessionCreateArgs,
-    format: Format,
-) -> Result<()> {
-    let body = build_session_create_body(args)?;
-    let resp = client
-        .post(SESSIONS_PATH, &body)
-        .context("failed to create agent session")?;
-    print_session(&resp, format);
-    Ok(())
-}
-
-pub fn list_sessions(
-    client: &ApiClient,
-    provider: Option<&str>,
-    state: Option<&str>,
-    limit: Option<usize>,
-    full: bool,
-    format: Format,
-) -> Result<()> {
-    let summary_limit = if full {
-        None
-    } else {
-        let limit = limit.unwrap_or(DEFAULT_SESSION_LIST_LIMIT);
-        if limit == 0 {
-            bail!("--limit must be greater than 0");
-        }
-        Some(limit)
-    };
-    let resp = client
-        .get(&sessions_path(provider, state, summary_limit))
-        .context("failed to list agent sessions")?;
-    print_sessions(&resp, format);
-    Ok(())
-}
-
-pub fn get_session(client: &ApiClient, id: &str, format: Format) -> Result<()> {
-    let resp = client
-        .get(&format!("{SESSIONS_PATH}/{id}"))
-        .with_context(|| format!("failed to get agent session {id}"))?;
-    print_session(&resp, format);
-    Ok(())
-}
-
-pub fn update_session(
-    client: &ApiClient,
-    args: &AgentSessionUpdateArgs,
-    format: Format,
-) -> Result<()> {
-    let body = build_session_update_body(args)?;
-    let resp = client
-        .patch(&format!("{SESSIONS_PATH}/{}", args.id), &body)
-        .with_context(|| format!("failed to update agent session {}", args.id))?;
-    print_session(&resp, format);
-    Ok(())
-}
-
-pub fn create_turn(client: &ApiClient, args: &AgentTurnCreateArgs, format: Format) -> Result<()> {
-    let body = build_turn_create_body(args)?;
-    let resp = client
-        .post(&format!("{SESSIONS_PATH}/{}/turns", args.session_id), &body)
-        .with_context(|| format!("failed to create agent turn in session {}", args.session_id))?;
-    print_turn(&resp, format);
-    Ok(())
-}
-
-pub fn list_turns(
-    client: &ApiClient,
-    session_id: &str,
-    status: Option<&str>,
-    format: Format,
-) -> Result<()> {
-    let resp = client
-        .get(&session_turns_path(session_id, status))
-        .with_context(|| format!("failed to list agent turns for session {session_id}"))?;
-    print_turns(&resp, format);
-    Ok(())
-}
-
-pub fn get_turn(client: &ApiClient, id: &str, format: Format) -> Result<()> {
-    let resp = client
-        .get(&format!("{TURNS_PATH}/{id}"))
-        .with_context(|| format!("failed to get agent turn {id}"))?;
-    print_turn(&resp, format);
-    Ok(())
-}
-
-pub fn transcript_turn(client: &ApiClient, id: &str, format: Format) -> Result<()> {
-    let turn = client
-        .get(&format!("{TURNS_PATH}/{id}"))
-        .with_context(|| format!("failed to get agent turn {id}"))?;
-    let event_values = list_turn_event_values(client, id)?;
-    match format {
-        Format::Json => output::print_json(&json!({
-            "turn": turn,
-            "events": event_values,
-        })),
-        Format::Table => {
-            let turn = decode_json::<AgentTurnInfo>(turn)?;
-            let events: Vec<AgentTurnEventInfo> = event_values
-                .into_iter()
-                .map(decode_json)
-                .collect::<Result<_>>()?;
-            render_turn_transcript(&turn, &events)?;
-        }
-    }
-    Ok(())
-}
-
-pub fn cancel_turn(
-    client: &ApiClient,
-    id: &str,
-    reason: Option<&str>,
-    format: Format,
-) -> Result<()> {
-    let body = match reason {
-        Some(reason) => json!({ "reason": reason }),
-        None => json!({}),
-    };
-    let resp = client
-        .post(&format!("{TURNS_PATH}/{id}/cancel"), &body)
-        .with_context(|| format!("failed to cancel agent turn {id}"))?;
-    print_turn(&resp, format);
-    Ok(())
-}
-
-fn cancel_turn_silent(client: &ApiClient, id: &str, reason: &str) -> Result<AgentTurnInfo> {
-    decode_json(
-        client
-            .post(
-                &format!("{TURNS_PATH}/{id}/cancel"),
-                &json!({ "reason": reason }),
-            )
-            .with_context(|| format!("failed to cancel agent turn {id}"))?,
-    )
-}
-
-pub fn list_turn_events(
-    client: &ApiClient,
-    args: &AgentTurnEventListArgs,
-    format: Format,
-) -> Result<()> {
-    let resp = client
-        .get(&turn_events_path(
-            &args.id, false, args.after, args.limit, None,
-        ))
-        .with_context(|| format!("failed to list events for agent turn {}", args.id))?;
-    print_turn_events(&resp, format);
-    Ok(())
-}
-
-pub fn stream_turn_events(client: &ApiClient, args: &AgentTurnEventStreamArgs) -> Result<()> {
-    let mut resp = client
-        .get_stream(&turn_events_path(
-            &args.id, true, args.after, args.limit, None,
-        ))
-        .with_context(|| format!("failed to stream events for agent turn {}", args.id))?;
-    let mut stdout = io::stdout().lock();
-    io::copy(&mut resp, &mut stdout).context("failed to read agent turn event stream")?;
-    Ok(())
-}
-
-pub fn run_interactive(client: &ApiClient, args: &AgentArgs) -> Result<()> {
-    let shell = AgentShell::connect(client, args)?;
-    run_shell_interactive(client, shell, args.messages.clone())
-}
-
-pub fn resume_interactive(client: &ApiClient, args: &AgentResumeArgs) -> Result<()> {
-    let shell = AgentShell::resume(client, args)?;
-    run_shell_interactive(client, shell, args.messages.clone())
-}
-
-fn run_shell_interactive(
-    client: &ApiClient,
-    mut shell: AgentShell,
-    initial_messages: Vec<String>,
-) -> Result<()> {
-    let session_id = shell.session.id.clone();
-    if tui::can_run() {
-        let result = tui::run_shell(client, shell, initial_messages);
-        if result.is_ok() {
-            print_resume_command(&session_id)?;
-        }
-        return result;
-    }
-
-    shell.print_banner()?;
-    let interrupts = InterruptState::install();
-    let mut input = InteractiveLineReader::with_history_namespace("agent")?;
-
-    if !initial_messages.is_empty() {
-        shell.submit_turn(client, initial_messages, &interrupts)?;
-    }
-
-    loop {
-        let Some(line) = prompt_agent_message(&mut input)? else {
-            print_resume_command(&session_id)?;
-            return Ok(());
-        };
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if handle_agent_slash_command(client, &mut shell, trimmed)? {
-            continue;
-        }
-        shell.submit_turn(client, vec![line], &interrupts)?;
-    }
-}
-
-fn print_resume_command(session_id: &str) -> Result<()> {
-    let mut stdout = io::stdout().lock();
-    writeln!(stdout, "Resume with: gestalt agent resume {session_id}")?;
-    Ok(())
-}
-
-fn handle_agent_slash_command(
-    client: &ApiClient,
-    shell: &mut AgentShell,
-    trimmed: &str,
-) -> Result<bool> {
-    let Some((command, args)) = parse_agent_slash_command(trimmed) else {
-        return Ok(false);
-    };
-    match command {
-        "help" => {
-            for line in agent_help_lines() {
-                eprintln!("{line}");
-            }
-        }
-        "session" => {
-            for line in agent_session_lines(shell) {
-                eprintln!("{line}");
-            }
-        }
-        "model" => {
-            for line in agent_model_lines(client, shell, args) {
-                eprintln!("{line}");
-            }
-        }
-        _ => return Ok(false),
-    }
-    Ok(true)
-}
-
-fn parse_agent_slash_command(input: &str) -> Option<(&str, &str)> {
-    let trimmed = input.trim();
-    let command = trimmed.strip_prefix('/')?;
-    let mut parts = command.splitn(2, char::is_whitespace);
-    let command = parts.next().unwrap_or("");
-    let args = parts.next().unwrap_or("");
-    Some((command, args.trim()))
-}
-
-fn agent_help_lines() -> Vec<String> {
-    vec![
-        "Commands:".to_string(),
-        "  /help     Show commands and keys.".to_string(),
-        "  /session  Show the active session id and resume command.".to_string(),
-        "  /model    Show current model and configured providers.".to_string(),
-        "  /model X  Use model X for future turns in this session.".to_string(),
-        "Keys:".to_string(),
-        "  Enter sends; busy turns queue the prompt.".to_string(),
-        "  Alt-Enter inserts a newline.".to_string(),
-        "  Up/Down recalls prompt history.".to_string(),
-        "  PgUp/PgDn scrolls the transcript.".to_string(),
-        "  Ctrl-C cancels, clears input, or exits.".to_string(),
-    ]
-}
-
-fn agent_tui_help_lines() -> Vec<String> {
-    let mut lines = agent_help_lines();
-    for line in &mut lines {
-        match line.as_str() {
-            "  Alt-Enter inserts a newline." => {
-                *line =
-                    "  Alt-Enter/Ctrl-J/Shift-Enter inserts a newline where supported.".to_string();
-            }
-            "  Up/Down recalls prompt history." => {
-                *line = "  Up/Down recalls input history.".to_string();
-            }
-            "  PgUp/PgDn scrolls the transcript." => {
-                *line = "  PgUp/PgDn scrolls the transcript by half a screen.".to_string();
-            }
-            _ => {}
-        }
-    }
-    let insert_at = lines
-        .iter()
-        .position(|line| line.contains("Ctrl-C cancels"))
-        .unwrap_or(lines.len());
-    lines.insert(
-        insert_at,
-        "  Ctrl-D exits when idle and input is empty.".to_string(),
-    );
-    lines.insert(insert_at, "  Ctrl-L redraws the terminal.".to_string());
-    lines.insert(
-        insert_at,
-        "  Ctrl-Home/Ctrl-End jumps to transcript start/end.".to_string(),
-    );
-    lines.insert(
-        insert_at,
-        "  Wheel/trackpad scrolls the transcript.".to_string(),
-    );
-    lines.insert(
-        insert_at,
-        "  Ctrl-O toggles compact/full tool details.".to_string(),
-    );
-    lines
-}
-
-fn agent_session_lines(shell: &AgentShell) -> Vec<String> {
-    vec![
-        format!("session {}", shell.session.id),
-        format!("resume command: gestalt agent resume {}", shell.session.id),
-    ]
-}
-
-fn agent_model_lines(client: &ApiClient, shell: &mut AgentShell, args: &str) -> Vec<String> {
-    let requested = args.trim();
-    if !requested.is_empty() {
-        shell.set_model_override(requested);
-        return vec![format!("model {requested} selected for future turns")];
-    }
-
-    let mut lines = vec![
-        format!("current provider: {}", shell.session.provider),
-        format!("current model: {}", shell.effective_model_label()),
-    ];
-    match list_agent_providers(client) {
-        Ok(providers) if providers.is_empty() => {
-            lines.push("configured providers: none".to_string());
-        }
-        Ok(providers) => {
-            lines.push("configured providers:".to_string());
-            for provider in providers {
-                let suffix = if provider.default { " (default)" } else { "" };
-                lines.push(format!("  {}{}", provider.name, suffix));
-            }
-        }
-        Err(err) => {
-            lines.push(format!("configured providers unavailable: {err}"));
-        }
-    }
-    lines
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentSessionInfo {
-    id: String,
-    provider: String,
-    #[serde(default)]
-    model: String,
-    #[serde(default)]
-    state: String,
-    #[serde(default)]
-    last_turn_at: String,
-    #[serde(default)]
-    created_at: String,
-    #[serde(default)]
-    updated_at: String,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentProviderListInfo {
-    #[serde(default)]
-    providers: Vec<AgentProviderInfo>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentProviderInfo {
-    name: String,
-    #[serde(default)]
-    default: bool,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentTurnInfo {
-    id: String,
-    #[serde(default)]
-    messages: Vec<AgentMessageInfo>,
-    #[serde(default)]
-    status: String,
-    #[serde(default)]
-    output_text: String,
-    #[serde(default)]
-    structured_output: Option<Value>,
-    #[serde(default)]
-    status_message: String,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentMessageInfo {
-    #[serde(default)]
-    role: String,
-    #[serde(default)]
-    text: String,
-    #[serde(default)]
-    parts: Vec<AgentMessagePartInfo>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentMessagePartInfo {
-    #[serde(default)]
-    text: String,
-    #[serde(default)]
-    json: Option<Value>,
-    #[serde(default)]
-    tool_call: Option<Value>,
-    #[serde(default)]
-    tool_result: Option<Value>,
-    #[serde(default)]
-    image_ref: Option<Value>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentTurnEventInfo {
-    #[serde(default)]
-    id: String,
-    #[serde(default)]
-    turn_id: String,
-    #[serde(default)]
-    seq: i64,
-    #[serde(rename = "type")]
-    event_type: String,
-    #[serde(default)]
-    source: String,
-    #[serde(default)]
-    visibility: String,
-    #[serde(default)]
-    data: Map<String, Value>,
-    #[serde(default, deserialize_with = "deserialize_turn_display")]
-    display: Option<AgentTurnDisplayInfo>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentTurnDisplayInfo {
-    #[serde(default)]
-    kind: String,
-    #[serde(default)]
-    phase: String,
-    #[serde(default)]
-    text: String,
-    #[serde(default)]
-    label: String,
-    #[serde(default, rename = "ref")]
-    display_ref: String,
-    #[allow(dead_code)]
-    #[serde(default)]
-    parent_ref: String,
-    #[serde(default)]
-    input: Option<Value>,
-    #[serde(default)]
-    output: Option<Value>,
-    #[serde(default)]
-    error: Option<Value>,
-    #[serde(default)]
-    action: String,
-    #[serde(default)]
-    format: String,
-    #[serde(default)]
-    language: String,
-}
-
-impl AgentTurnDisplayInfo {
-    fn from_value(value: Value) -> Option<Self> {
-        let Value::Object(mut data) = value else {
-            return None;
-        };
-        Some(Self {
-            kind: take_string_field(&mut data, "kind"),
-            phase: take_string_field(&mut data, "phase"),
-            text: take_string_field(&mut data, "text"),
-            label: take_string_field(&mut data, "label"),
-            display_ref: take_string_field(&mut data, "ref"),
-            parent_ref: take_string_field(&mut data, "parentRef"),
-            input: data.remove("input"),
-            output: data.remove("output"),
-            error: data.remove("error"),
-            action: take_string_field(&mut data, "action"),
-            format: take_string_field(&mut data, "format"),
-            language: take_string_field(&mut data, "language"),
-        })
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AgentInteractionInfo {
-    id: String,
-    #[serde(rename = "type")]
-    interaction_type: String,
-    #[serde(default)]
-    state: String,
-    #[serde(default)]
-    title: String,
-    #[serde(default)]
-    prompt: String,
-    #[serde(default)]
-    request: Map<String, Value>,
-}
-
-struct AgentShell {
-    session: AgentSessionInfo,
-    model_override: Option<String>,
-    system_messages: Vec<String>,
-    tools: Vec<AgentToolArg>,
-    applied_system_messages: bool,
-}
-
-impl AgentShell {
-    fn connect(client: &ApiClient, args: &AgentArgs) -> Result<Self> {
-        let session_args = AgentSessionCreateArgs {
-            provider: args.provider.clone(),
-            model: args.model.clone(),
-            client_ref: None,
-            idempotency_key: None,
-            input: None,
-        };
-        let session = create_session_info(client, &session_args)?;
-
-        Ok(Self {
-            session,
-            model_override: args.model.clone(),
-            system_messages: args.system.clone(),
-            tools: args.tools.clone(),
-            applied_system_messages: false,
-        })
-    }
-
-    fn resume(client: &ApiClient, args: &AgentResumeArgs) -> Result<Self> {
-        let session = match args.session_id.as_deref() {
-            Some(session_id) => get_session_info(client, session_id)?,
-            None => resume_latest_session_info(client, args.provider.as_deref())?,
-        };
-
-        Ok(Self {
-            session,
-            model_override: args.model.clone(),
-            system_messages: args.system.clone(),
-            tools: args.tools.clone(),
-            applied_system_messages: false,
-        })
-    }
-
-    fn print_banner(&self) -> Result<()> {
-        let mut stderr = io::stderr().lock();
-        writeln!(
-            stderr,
-            "Session {} [{} / {}]",
-            self.session.id,
-            self.session.provider,
-            self.effective_model_label()
-        )?;
-        writeln!(stderr, "Press Ctrl-C or send EOF to exit.")?;
-        Ok(())
-    }
-
-    fn effective_model_label(&self) -> &str {
-        if let Some(model) = self.model_override.as_deref()
-            && !model.trim().is_empty()
-        {
-            return model;
-        }
-        if self.session.model.trim().is_empty() {
-            "<unspecified>"
-        } else {
-            self.session.model.as_str()
-        }
-    }
-
-    fn set_model_override(&mut self, model: &str) {
-        self.model_override = Some(model.trim().to_string());
-    }
-
-    fn submit_turn(
-        &mut self,
-        client: &ApiClient,
-        messages: Vec<String>,
-        interrupts: &InterruptState,
-    ) -> Result<()> {
-        let system_messages = if self.applied_system_messages {
-            Vec::new()
-        } else {
-            self.system_messages.clone()
-        };
-        let turn_args = AgentTurnCreateArgs {
-            session_id: self.session.id.clone(),
-            model: self.model_override.clone(),
-            system: system_messages,
-            messages,
-            tools: self.tools.clone(),
-            idempotency_key: None,
-            input: None,
-        };
-        let turn = create_turn_info(client, &turn_args)?;
-        self.applied_system_messages = true;
-        drive_turn(client, &turn, interrupts)?;
-        Ok(())
-    }
-}
-
-fn drive_turn(client: &ApiClient, turn: &AgentTurnInfo, interrupts: &InterruptState) -> Result<()> {
-    let mut renderer = AgentTurnRenderer::new();
-    let _cancel_guard = TurnCancelGuard::spawn(client, &turn.id, interrupts);
-    loop {
-        stream_turn_events_until_blocked_or_terminal(client, &turn.id, &mut renderer)?;
-        let latest = get_turn_info(client, &turn.id)?;
-        renderer.finish_turn(&latest)?;
-
-        match latest.status.as_str() {
-            "waiting_for_input" => {
-                let interactions = list_interactions_info(client, &turn.id)?;
-                let pending: Vec<_> = interactions
-                    .into_iter()
-                    .filter(|interaction| interaction.state == "pending")
-                    .collect();
-                if pending.is_empty() {
-                    bail!(
-                        "agent turn {} is waiting for input without a pending interaction",
-                        latest.id
-                    );
-                }
-                for interaction in pending {
-                    let prompt_interrupt_count = interrupts.count();
-                    let resolution = match prompt_interaction_resolution(&interaction) {
-                        Ok(resolution) => resolution,
-                        Err(_) if interrupts.count() > prompt_interrupt_count => {
-                            let _ = cancel_turn_silent(client, &turn.id, INTERRUPT_CANCEL_REASON);
-                            return Ok(());
-                        }
-                        Err(err) => return Err(err),
-                    };
-                    if interrupts.count() > prompt_interrupt_count {
-                        let _ = cancel_turn_silent(client, &turn.id, INTERRUPT_CANCEL_REASON);
-                        return Ok(());
-                    }
-                    resolve_interaction_info(client, &turn.id, &interaction.id, resolution)?;
-                }
-            }
-            "pending" | "running" => thread::sleep(EVENT_POLL_INTERVAL),
-            "succeeded" | "failed" | "canceled" => return Ok(()),
-            other => bail!("agent turn {} has unsupported status {}", latest.id, other),
-        }
-    }
-}
-
-#[derive(Clone)]
-struct InterruptState {
-    count: Arc<AtomicU64>,
-}
-
-impl InterruptState {
-    fn install() -> Self {
-        let count = Arc::new(AtomicU64::new(0));
-        let handler_count = Arc::clone(&count);
-        if let Err(err) = ctrlc::set_handler(move || {
-            handler_count.fetch_add(1, Ordering::SeqCst);
-        }) {
-            eprintln!("warning: failed to install Ctrl-C handler: {err}");
-        }
-        Self { count }
-    }
-
-    fn count(&self) -> u64 {
-        self.count.load(Ordering::SeqCst)
-    }
-}
-
-struct TurnCancelGuard {
-    active: Arc<AtomicBool>,
-    handle: Option<thread::JoinHandle<()>>,
-}
-
-impl TurnCancelGuard {
-    fn spawn(client: &ApiClient, turn_id: &str, interrupts: &InterruptState) -> Self {
-        let active = Arc::new(AtomicBool::new(true));
-        let thread_active = Arc::clone(&active);
-        let client = client.clone();
-        let turn_id = turn_id.to_string();
-        let interrupts = interrupts.clone();
-        let baseline = interrupts.count();
-        let handle = thread::spawn(move || {
-            while thread_active.load(Ordering::SeqCst) {
-                if interrupts.count() > baseline {
-                    let _ = cancel_turn_silent(&client, &turn_id, INTERRUPT_CANCEL_REASON);
-                    return;
-                }
-                thread::sleep(Duration::from_millis(100));
-            }
-        });
-        Self {
-            active,
-            handle: Some(handle),
-        }
-    }
-}
-
-impl Drop for TurnCancelGuard {
-    fn drop(&mut self) {
-        self.active.store(false, Ordering::SeqCst);
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.join();
-        }
-    }
-}
-
-struct AgentTurnRenderer {
-    after_seq: u64,
-    assistant_line_open: bool,
-    saw_assistant_output: bool,
-    saw_structured_output: bool,
-    delta_buffer: String,
-    use_color: bool,
-}
-
-impl AgentTurnRenderer {
-    fn new() -> Self {
-        Self {
-            after_seq: 0,
-            assistant_line_open: false,
-            saw_assistant_output: false,
-            saw_structured_output: false,
-            delta_buffer: String::new(),
-            use_color: io::stdout().is_terminal(),
-        }
-    }
-
-    fn after_seq(&self) -> u64 {
-        self.after_seq
-    }
-
-    fn render_messages(&mut self, messages: &[AgentMessageInfo]) -> Result<()> {
-        self.finish_assistant_line();
-        for message in messages {
-            let Some(label) = message_label(message) else {
-                continue;
-            };
-            let Some(text) = message_text(message) else {
-                continue;
-            };
-            println!("{} {text}", self.label(&label));
-        }
-        Ok(())
-    }
-
-    fn render_events(&mut self, events: &[AgentTurnEventInfo]) -> Result<()> {
-        for event in events {
-            if event.seq > 0 {
-                self.after_seq = self.after_seq.max(event.seq as u64);
-            }
-            if self.render_display_event(event)? {
-                continue;
-            }
-            match event.event_type.as_str() {
-                "agent.message.delta" | "assistant.delta" => {
-                    if let Some(text) = string_any_field(&event.data, &["text", "delta", "content"])
-                    {
-                        self.start_assistant_line()?;
-                        print!("{text}");
-                        io::stdout().flush().context("failed to flush stdout")?;
-                        self.saw_assistant_output = true;
-                        self.delta_buffer.push_str(&text);
-                    }
-                }
-                "assistant.completed" => {
-                    let text = string_field(&event.data, "text");
-                    if self.assistant_line_open {
-                        if let Some(text) = text.as_deref() {
-                            if self.delta_buffer.is_empty() {
-                                print!("{text}");
-                            } else if let Some(suffix) = text.strip_prefix(&self.delta_buffer) {
-                                print!("{suffix}");
-                            }
-                        }
-                        println!();
-                        self.assistant_line_open = false;
-                    } else if let Some(text) = text {
-                        println!("{} {text}", self.label("assistant>"));
-                        self.saw_assistant_output = true;
-                    }
-                    self.delta_buffer.clear();
-                }
-                "turn.started" => {
-                    self.finish_assistant_line();
-                    match string_any_field(&event.data, &["status", "state"]) {
-                        Some(status) => println!("{} started ({status})", self.label("turn>")),
-                        None => println!("{} started", self.label("turn>")),
-                    }
-                }
-                "tool.started" => {
-                    self.finish_assistant_line();
-                    let tool = string_any_field(
-                        &event.data,
-                        &[
-                            "tool_name",
-                            "toolName",
-                            "name",
-                            "operation",
-                            "tool_id",
-                            "toolId",
-                        ],
-                    )
-                    .unwrap_or_else(|| "tool".to_string());
-                    print!("{} {tool} started", self.label("tool>"));
-                    if let Some(input) =
-                        value_any_field(&event.data, &["arguments", "input", "request"])
-                    {
-                        print!(" {}", compact_json(input)?);
-                    }
-                    println!();
-                }
-                "tool.completed" => {
-                    self.finish_assistant_line();
-                    let tool = string_any_field(
-                        &event.data,
-                        &[
-                            "tool_name",
-                            "toolName",
-                            "name",
-                            "operation",
-                            "tool_id",
-                            "toolId",
-                        ],
-                    )
-                    .unwrap_or_else(|| "tool".to_string());
-                    let status =
-                        string_any_field(&event.data, &["status", "state"]).or_else(|| {
-                            number_any_field(&event.data, &["status", "statusCode"])
-                                .map(|status| status.to_string())
-                        });
-                    match status {
-                        Some(status) => {
-                            print!("{} {tool} completed ({status})", self.label("tool>"))
-                        }
-                        None => print!("{} {tool} completed", self.label("tool>")),
-                    }
-                    if let Some(error) = string_field(&event.data, "error") {
-                        print!(": {error}");
-                    } else if let Some(output) =
-                        value_any_field(&event.data, &["output", "result", "body"])
-                    {
-                        print!(" {}", compact_json(output)?);
-                    }
-                    println!();
-                }
-                "interaction.requested" => {
-                    self.finish_assistant_line();
-                    let interaction_id =
-                        string_any_field(&event.data, &["interaction_id", "interactionId"])
-                            .unwrap_or_else(|| "interaction".to_string());
-                    println!(
-                        "{} requested ({interaction_id})",
-                        self.label("interaction>")
-                    );
-                }
-                "interaction.resolved" => {
-                    self.finish_assistant_line();
-                    let interaction_id =
-                        string_any_field(&event.data, &["interaction_id", "interactionId"])
-                            .unwrap_or_else(|| "interaction".to_string());
-                    println!("{} resolved ({interaction_id})", self.label("interaction>"));
-                }
-                "turn.failed" => {
-                    self.finish_assistant_line();
-                    if let Some(message) = string_field(&event.data, "error") {
-                        println!("{} failed: {message}", self.label("turn>"));
-                    }
-                }
-                "turn.canceled" => {
-                    self.finish_assistant_line();
-                    if let Some(reason) = string_field(&event.data, "reason") {
-                        println!("{} canceled: {reason}", self.label("turn>"));
-                    }
-                }
-                "turn.completed" => {}
-                _ if event.visibility == "private" => {}
-                _ => self.render_generic_event(event)?,
-            }
-        }
-        Ok(())
-    }
-
-    fn finish_turn(&mut self, turn: &AgentTurnInfo) -> Result<()> {
-        self.finish_assistant_line();
-        match turn.status.as_str() {
-            "succeeded" if !self.saw_assistant_output && !turn.output_text.is_empty() => {
-                println!("{} {}", self.label("assistant>"), turn.output_text);
-                self.saw_assistant_output = true;
-            }
-            "failed" if !turn.status_message.is_empty() => {
-                println!("{} failed: {}", self.label("turn>"), turn.status_message);
-            }
-            "canceled" if !turn.status_message.is_empty() => {
-                println!("{} canceled: {}", self.label("turn>"), turn.status_message);
-            }
-            _ => {}
-        }
-        if !self.saw_structured_output
-            && let Some(structured_output) = turn.structured_output.as_ref()
-        {
-            println!(
-                "{} {}",
-                self.label("structured>"),
-                pretty_json(structured_output)?
-            );
-            self.saw_structured_output = true;
-        }
-        self.delta_buffer.clear();
-        Ok(())
-    }
-
-    fn start_assistant_line(&mut self) -> Result<()> {
-        if !self.assistant_line_open {
-            print!("{} ", self.label("assistant>"));
-            io::stdout().flush().context("failed to flush stdout")?;
-            self.assistant_line_open = true;
-        }
-        Ok(())
-    }
-
-    fn finish_assistant_line(&mut self) {
-        if self.assistant_line_open {
-            println!();
-            self.assistant_line_open = false;
-        }
-    }
-
-    fn render_display_event(&mut self, event: &AgentTurnEventInfo) -> Result<bool> {
-        let Some(display) = turn_event_display(event) else {
-            return Ok(false);
-        };
-        match display.kind.trim() {
-            "text" => self.render_display_text(display, "assistant>"),
-            "reasoning" => self.render_display_text(display, "reasoning>"),
-            "tool" => self.render_display_tool(event, display),
-            "interaction" => Ok(self.render_display_interaction(event, display)),
-            "status" => Ok(self.render_display_status(display)),
-            "error" => Ok(self.render_display_error(display)),
-            _ => Ok(false),
-        }
-    }
-
-    fn render_display_text(&mut self, display: &AgentTurnDisplayInfo, label: &str) -> Result<bool> {
-        let raw_text = display_text(display);
-        let rendered_text = if label == "assistant>" {
-            rendered_display_text(display)
-        } else {
-            raw_text.map(ToString::to_string)
-        };
-        if label == "assistant>" {
-            match display.phase.trim() {
-                "delta" => {
-                    if let Some(text) = raw_text {
-                        self.start_assistant_line()?;
-                        print!("{text}");
-                        io::stdout().flush().context("failed to flush stdout")?;
-                        self.saw_assistant_output = true;
-                        self.delta_buffer.push_str(text);
-                        return Ok(true);
-                    }
-                }
-                "completed" => {
-                    if self.assistant_line_open {
-                        let Some(text) = raw_text else {
-                            return Ok(false);
-                        };
-                        if self.delta_buffer.is_empty() {
-                            print!("{text}");
-                        } else if let Some(suffix) = text.strip_prefix(&self.delta_buffer) {
-                            print!("{suffix}");
-                        }
-                        println!();
-                        self.assistant_line_open = false;
-                        self.delta_buffer.clear();
-                        return Ok(true);
-                    } else if let Some(text) = rendered_text.as_deref() {
-                        println!("{} {text}", self.label(label));
-                        self.saw_assistant_output = true;
-                        self.delta_buffer.clear();
-                        return Ok(true);
-                    }
-                }
-                _ if rendered_text.is_some() => {
-                    self.finish_assistant_line();
-                    let text = rendered_text.as_deref().expect("checked is_some");
-                    println!("{} {text}", self.label(label));
-                    self.saw_assistant_output = true;
-                    return Ok(true);
-                }
-                _ => {}
-            }
-            return Ok(false);
-        }
-
-        self.finish_assistant_line();
-        if let Some(text) = rendered_text.as_deref() {
-            println!("{} {text}", self.label(label));
-            return Ok(true);
-        }
-        Ok(false)
-    }
-
-    fn render_display_tool(
-        &mut self,
-        event: &AgentTurnEventInfo,
-        display: &AgentTurnDisplayInfo,
-    ) -> Result<bool> {
-        self.finish_assistant_line();
-        let tool = display_tool_label(event, display);
-        let action = display_action(display);
-        match display.phase.trim() {
-            "started" => {
-                print!(
-                    "{} {}",
-                    self.label("tool>"),
-                    action
-                        .map(|action| format!("{action} {tool}"))
-                        .unwrap_or_else(|| format!("{tool} started"))
-                );
-                if let Some(input) = display_tool_input(event, display) {
-                    print!(" {}", compact_json(input)?);
-                }
-                println!();
-            }
-            "completed" => {
-                print!(
-                    "{} {}",
-                    self.label("tool>"),
-                    action
-                        .map(|action| format!("{action} {tool}"))
-                        .unwrap_or_else(|| format!("{tool} completed"))
-                );
-                if let Some(status) = display_status(event, display) {
-                    print!(" ({status})");
-                }
-                if let Some(error) = display_tool_error(event, display) {
-                    print!(": {}", display_value_text(error)?);
-                } else if let Some(output) = display_tool_output(event, display) {
-                    print!(" {}", compact_json(output)?);
-                }
-                println!();
-            }
-            "failed" => {
-                print!(
-                    "{} {}",
-                    self.label("tool>"),
-                    action
-                        .map(|action| format!("{action} {tool}"))
-                        .unwrap_or_else(|| format!("{tool} failed"))
-                );
-                if let Some(status) = display_status(event, display) {
-                    print!(" ({status})");
-                }
-                if let Some(error) = display_tool_error(event, display) {
-                    print!(": {}", display_value_text(error)?);
-                }
-                println!();
-            }
-            "progress" => match (action, display_text(display)) {
-                (Some(action), Some(text)) => {
-                    println!("{} {action} {tool}: {text}", self.label("tool>"));
-                }
-                (Some(action), None) => {
-                    println!("{} {action} {tool}", self.label("tool>"));
-                }
-                (None, Some(text)) => {
-                    println!("{} {tool} {text}", self.label("tool>"));
-                }
-                (None, None) => {
-                    println!("{} {tool} progress", self.label("tool>"));
-                }
-            },
-            _ => return Ok(false),
-        }
-        Ok(true)
-    }
-
-    fn render_display_interaction(
-        &mut self,
-        _event: &AgentTurnEventInfo,
-        display: &AgentTurnDisplayInfo,
-    ) -> bool {
-        self.finish_assistant_line();
-        let interaction_ref = display_ref(display)
-            .map(ToString::to_string)
-            .unwrap_or_else(|| "interaction".to_string());
-        match display.phase.trim() {
-            "requested" => println!(
-                "{} requested ({interaction_ref})",
-                self.label("interaction>")
-            ),
-            "resolved" => println!(
-                "{} resolved ({interaction_ref})",
-                self.label("interaction>")
-            ),
-            _ => return false,
-        }
-        true
-    }
-
-    fn render_display_status(&mut self, display: &AgentTurnDisplayInfo) -> bool {
-        self.finish_assistant_line();
-        let text = display_text(display);
-        match display.phase.trim() {
-            "started" => match text {
-                Some(text) => println!("{} started ({text})", self.label("turn>")),
-                None => println!("{} started", self.label("turn>")),
-            },
-            "canceled" => match text {
-                Some(text) => println!("{} canceled: {text}", self.label("turn>")),
-                None => println!("{} canceled", self.label("turn>")),
-            },
-            "completed" => {
-                if let Some(text) = text {
-                    println!("{} completed ({text})", self.label("turn>"));
-                }
-            }
-            "progress" => {
-                if let Some(text) = text {
-                    println!("{} {text}", self.label("turn>"));
-                }
-            }
-            _ => return false,
-        }
-        true
-    }
-
-    fn render_display_error(&mut self, display: &AgentTurnDisplayInfo) -> bool {
-        self.finish_assistant_line();
-        let label = if display_label(display) == Some("turn") {
-            "turn>"
-        } else {
-            "error>"
-        };
-        let text = display_text(display).map(ToString::to_string).or_else(|| {
-            display
-                .error
-                .as_ref()
-                .and_then(|value| display_value_text(value).ok())
-        });
-        let Some(text) = text else {
-            return false;
-        };
-        match display.phase.trim() {
-            "failed" => println!("{} failed: {text}", self.label(label)),
-            _ => println!("{} {text}", self.label(label)),
-        }
-        true
-    }
-
-    fn render_generic_event(&mut self, event: &AgentTurnEventInfo) -> Result<()> {
-        self.finish_assistant_line();
-        let mut fields = Vec::new();
-        if !event.id.is_empty() {
-            fields.push(format!("id={}", event.id));
-        }
-        if !event.source.is_empty() {
-            fields.push(format!("source={}", event.source));
-        }
-        if !event.turn_id.is_empty() {
-            fields.push(format!("turn={}", event.turn_id));
-        }
-        let suffix = if fields.is_empty() {
-            String::new()
-        } else {
-            format!(" ({})", fields.join(" "))
-        };
-        if event.data.is_empty() {
-            println!("{} {}{}", self.label("event>"), event.event_type, suffix);
-        } else {
-            println!(
-                "{} {}{} {}",
-                self.label("event>"),
-                event.event_type,
-                suffix,
-                compact_json(&Value::Object(event.data.clone()))?
-            );
-        }
-        Ok(())
-    }
-
-    fn label(&self, value: &str) -> String {
-        if self.use_color {
-            value.bold().cyan().to_string()
-        } else {
-            value.to_string()
-        }
-    }
-}
-
-fn prompt_agent_message(input: &mut InteractiveLineReader) -> Result<Option<String>> {
-    let mut lines = Vec::new();
-    let mut prompt = "agent> ";
-    loop {
-        match input.read_line(prompt)? {
-            PromptLine::Line(mut line) => {
-                let continued = has_trailing_continuation(&line);
-                if continued {
-                    line.pop();
-                }
-                lines.push(line);
-                if !continued {
-                    return Ok(Some(lines.join("\n")));
-                }
-                prompt = "...> ";
-            }
-            PromptLine::Interrupted => {
-                eprintln!("^C");
-                return Ok(None);
-            }
-            PromptLine::Eof => return Ok(None),
-        }
-    }
-}
-
-fn has_trailing_continuation(line: &str) -> bool {
-    line.chars().rev().take_while(|ch| *ch == '\\').count() % 2 == 1
-}
-
-fn prompt_interaction_resolution(interaction: &AgentInteractionInfo) -> Result<Map<String, Value>> {
-    let mut stderr = io::stderr().lock();
-    writeln!(stderr)?;
-    writeln!(
-        stderr,
-        "Interaction {} [{}]",
-        interaction.id, interaction.interaction_type
-    )?;
-    if !interaction.title.is_empty() {
-        writeln!(stderr, "{}", interaction.title)?;
-    }
-    if !interaction.prompt.is_empty() {
-        writeln!(stderr, "{}", interaction.prompt)?;
-    }
-    if !interaction.request.is_empty() {
-        writeln!(
-            stderr,
-            "Request: {}",
-            serde_json::to_string(&interaction.request)
-                .context("failed to encode interaction request")?
-        )?;
-    }
-    drop(stderr);
-
-    match interaction.interaction_type.as_str() {
-        "approval" => {
-            let approved = prompt_confirm("Approve?", true)?;
-            Ok(Map::from_iter([(
-                "approved".to_string(),
-                Value::Bool(approved),
-            )]))
-        }
-        "clarification" | "input" => {
-            let default = interaction
-                .request
-                .get("default")
-                .and_then(Value::as_str)
-                .map(ToString::to_string);
-            let required = interaction
-                .request
-                .get("required")
-                .and_then(Value::as_bool)
-                .unwrap_or(true);
-            let secret = interaction
-                .request
-                .get("secret")
-                .and_then(Value::as_bool)
-                .unwrap_or(false);
-            let label = if interaction.title.is_empty() {
-                "Response".to_string()
-            } else {
-                interaction.title.clone()
-            };
-            let description = if interaction.prompt.is_empty() {
-                None
-            } else {
-                Some(interaction.prompt.clone())
-            };
-            let response = prompt_input(&InputPrompt {
-                label,
-                description,
-                default,
-                required,
-                secret,
-            })?;
-            Ok(Map::from_iter([(
-                "response".to_string(),
-                Value::String(response),
-            )]))
-        }
-        other => bail!("unsupported agent interaction type {other}"),
-    }
-}
-
-fn stream_turn_events_until_blocked_or_terminal(
-    client: &ApiClient,
-    turn_id: &str,
-    renderer: &mut AgentTurnRenderer,
-) -> Result<()> {
-    let after_seq = renderer.after_seq();
-    stream_turn_event_frames(client, turn_id, after_seq, |event| {
-        renderer.render_events(&[event])
-    })
-}
-
-fn stream_turn_event_frames<F>(
-    client: &ApiClient,
-    turn_id: &str,
-    after_seq: u64,
-    mut handle_event: F,
-) -> Result<()>
-where
-    F: FnMut(AgentTurnEventInfo) -> Result<()>,
-{
-    let resp = client
-        .get_stream(&turn_events_path(
-            turn_id,
-            true,
-            Some(after_seq),
-            Some(DEFAULT_EVENT_PAGE_SIZE),
-            Some(EVENT_STREAM_UNTIL_BLOCKED_OR_TERMINAL),
-        ))
-        .with_context(|| format!("failed to stream events for agent turn {turn_id}"))?;
-    let mut reader = BufReader::new(resp);
-    let mut line = String::new();
-    let mut decoder = SseEventDecoder::default();
-
-    loop {
-        line.clear();
-        let read = reader
-            .read_line(&mut line)
-            .context("failed to read agent turn event stream")?;
-        if read == 0 {
-            if let Some(event) = decoder.finish()? {
-                handle_event(event)?;
-            }
-            return Ok(());
-        }
-
-        if let Some(event) = decoder.push_line(&line)? {
-            handle_event(event)?;
-        }
-    }
-}
-
-#[derive(Default)]
-struct SseEventDecoder {
-    data: String,
-}
-
-impl SseEventDecoder {
-    fn push_line(&mut self, line: &str) -> Result<Option<AgentTurnEventInfo>> {
-        let trimmed = line.trim_end_matches(['\r', '\n']);
-        if trimmed.is_empty() {
-            return self.finish();
-        }
-
-        if let Some(value) = trimmed.strip_prefix("data:") {
-            if !self.data.is_empty() {
-                self.data.push('\n');
-            }
-            self.data.push_str(value.strip_prefix(' ').unwrap_or(value));
-        }
-
-        Ok(None)
-    }
-
-    fn finish(&mut self) -> Result<Option<AgentTurnEventInfo>> {
-        if self.data.is_empty() {
-            return Ok(None);
-        }
-        let raw = std::mem::take(&mut self.data);
-        serde_json::from_str(&raw)
-            .with_context(|| format!("failed to decode agent turn event stream frame: {raw}"))
-            .map(Some)
-    }
-}
-
-fn create_session_info(
-    client: &ApiClient,
-    args: &AgentSessionCreateArgs,
-) -> Result<AgentSessionInfo> {
-    let body = build_session_create_body(args)?;
-    decode_json(
-        client
-            .post(SESSIONS_PATH, &body)
-            .context("failed to create agent session")?,
-    )
-}
-
-fn get_session_info(client: &ApiClient, id: &str) -> Result<AgentSessionInfo> {
-    decode_json(
-        client
-            .get(&format!("{SESSIONS_PATH}/{id}"))
-            .with_context(|| format!("failed to get agent session {id}"))?,
-    )
-}
-
-fn list_agent_providers(client: &ApiClient) -> Result<Vec<AgentProviderInfo>> {
-    let resp: AgentProviderListInfo = decode_json(
-        client
-            .get(PROVIDERS_PATH)
-            .context("failed to list agent providers")?,
-    )?;
-    Ok(resp.providers)
-}
-
-fn resume_latest_session_info(
-    client: &ApiClient,
-    provider: Option<&str>,
-) -> Result<AgentSessionInfo> {
-    let sessions: Vec<AgentSessionInfo> = decode_json(
-        client
-            .get(&sessions_path(
-                provider,
-                Some("active"),
-                Some(DEFAULT_SESSION_LIST_LIMIT),
-            ))
-            .context("failed to list active agent sessions")?,
-    )?;
-    sessions
-        .into_iter()
-        .filter(|session| session.state.is_empty() || session.state == "active")
-        .max_by(compare_sessions_for_resume)
-        .ok_or_else(|| match provider {
-            Some(provider) => anyhow::anyhow!(
-                "no active agent sessions found for provider {provider}; use `gestalt agent` to create one"
-            ),
-            None => anyhow::anyhow!(
-                "no active agent sessions found; use `gestalt agent` to create one"
-            ),
-        })
-}
-
-fn compare_sessions_for_resume(a: &AgentSessionInfo, b: &AgentSessionInfo) -> std::cmp::Ordering {
-    compare_session_time_field(&a.last_turn_at, &b.last_turn_at)
-        .then_with(|| compare_session_time_field(&a.updated_at, &b.updated_at))
-        .then_with(|| compare_session_time_field(&a.created_at, &b.created_at))
-        .then_with(|| a.id.cmp(&b.id))
-}
-
-fn compare_session_time_field(a: &str, b: &str) -> std::cmp::Ordering {
-    match (parse_session_time(a), parse_session_time(b)) {
-        (Some(a), Some(b)) => a.cmp(&b),
-        (Some(_), None) => std::cmp::Ordering::Greater,
-        (None, Some(_)) => std::cmp::Ordering::Less,
-        (None, None) => a.cmp(b),
-    }
-}
-
-fn parse_session_time(value: &str) -> Option<OffsetDateTime> {
-    OffsetDateTime::parse(value, &Rfc3339).ok()
-}
-
-fn create_turn_info(client: &ApiClient, args: &AgentTurnCreateArgs) -> Result<AgentTurnInfo> {
-    let body = build_turn_create_body(args)?;
-    decode_json(
-        client
-            .post(&format!("{SESSIONS_PATH}/{}/turns", args.session_id), &body)
-            .with_context(|| {
-                format!("failed to create agent turn in session {}", args.session_id)
-            })?,
-    )
-}
-
-fn get_turn_info(client: &ApiClient, id: &str) -> Result<AgentTurnInfo> {
-    decode_json(
-        client
-            .get(&format!("{TURNS_PATH}/{id}"))
-            .with_context(|| format!("failed to get agent turn {id}"))?,
-    )
-}
-
-fn list_interactions_info(client: &ApiClient, turn_id: &str) -> Result<Vec<AgentInteractionInfo>> {
-    decode_json(
-        client
-            .get(&format!("{TURNS_PATH}/{turn_id}/interactions"))
-            .with_context(|| format!("failed to list interactions for agent turn {turn_id}"))?,
-    )
-}
-
-fn resolve_interaction_info(
-    client: &ApiClient,
-    turn_id: &str,
-    interaction_id: &str,
-    resolution: Map<String, Value>,
-) -> Result<AgentInteractionInfo> {
-    decode_json(
-        client
-            .post(
-                &format!("{TURNS_PATH}/{turn_id}/interactions/{interaction_id}/resolve"),
-                &json!({ "resolution": resolution }),
-            )
-            .with_context(|| format!("failed to resolve interaction {interaction_id}"))?,
-    )
-}
-
-fn list_turn_event_values(client: &ApiClient, turn_id: &str) -> Result<Vec<Value>> {
-    let mut events = Vec::new();
-    let mut after_seq = 0u64;
-    loop {
-        let page: Vec<Value> = decode_json(
-            client
-                .get(&turn_events_path(
-                    turn_id,
-                    false,
-                    Some(after_seq),
-                    Some(DEFAULT_EVENT_PAGE_SIZE),
-                    None,
-                ))
-                .with_context(|| format!("failed to list events for agent turn {turn_id}"))?,
-        )?;
-        if page.is_empty() {
-            return Ok(events);
-        }
-
-        let next_after_seq = page
-            .iter()
-            .filter_map(|event| event.get("seq").and_then(Value::as_u64))
-            .max()
-            .unwrap_or(after_seq);
-        events.extend(page);
-        if next_after_seq <= after_seq {
-            return Ok(events);
-        }
-        after_seq = next_after_seq;
-    }
-}
-
-fn render_turn_transcript(turn: &AgentTurnInfo, events: &[AgentTurnEventInfo]) -> Result<()> {
-    let mut renderer = AgentTurnRenderer::new();
-    renderer.render_messages(&turn.messages)?;
-    renderer.render_events(events)?;
-    renderer.finish_turn(turn)
-}
-
-fn decode_json<T>(value: Value) -> Result<T>
-where
-    T: DeserializeOwned,
-{
-    serde_json::from_value(value).context("failed to decode agent response")
-}
-
-fn deserialize_turn_display<'de, D>(
-    deserializer: D,
-) -> std::result::Result<Option<AgentTurnDisplayInfo>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value = Option::<Value>::deserialize(deserializer)?;
-    Ok(value.and_then(AgentTurnDisplayInfo::from_value))
-}
-
-fn take_string_field(data: &mut Map<String, Value>, key: &str) -> String {
-    match data.remove(key) {
-        Some(Value::String(value)) => value,
-        _ => String::new(),
-    }
-}
-
-fn string_field(data: &Map<String, Value>, key: &str) -> Option<String> {
-    data.get(key)
-        .and_then(Value::as_str)
-        .map(ToString::to_string)
-}
-
-fn string_any_field(data: &Map<String, Value>, keys: &[&str]) -> Option<String> {
-    keys.iter().find_map(|key| string_field(data, key))
-}
-
-fn value_any_field<'a>(data: &'a Map<String, Value>, keys: &[&str]) -> Option<&'a Value> {
-    keys.iter().find_map(|key| data.get(*key))
-}
-
-fn number_any_field(data: &Map<String, Value>, keys: &[&str]) -> Option<i64> {
-    keys.iter().find_map(|key| data.get(*key)?.as_i64())
-}
-
-fn turn_event_display(event: &AgentTurnEventInfo) -> Option<&AgentTurnDisplayInfo> {
-    let display = event.display.as_ref()?;
-    if display.kind.trim().is_empty() {
-        return None;
-    }
-    if event.visibility == "private" && !known_turn_event_type(&event.event_type) {
-        return None;
-    }
-    Some(display)
-}
-
-fn known_turn_event_type(event_type: &str) -> bool {
-    matches!(
-        event_type,
-        "agent.message.delta"
-            | "assistant.delta"
-            | "assistant.completed"
-            | "turn.started"
-            | "turn.completed"
-            | "turn.failed"
-            | "turn.canceled"
-            | "tool.started"
-            | "tool.completed"
-            | "tool.failed"
-            | "interaction.requested"
-            | "interaction.resolved"
-    )
-}
-
-fn message_label(message: &AgentMessageInfo) -> Option<String> {
-    let role = non_empty_str(&message.role)?;
-    Some(match role {
-        "user" => "you>".to_string(),
-        "system" => "system>".to_string(),
-        "assistant" => "assistant>".to_string(),
-        "tool" => "tool>".to_string(),
-        other => format!("{other}>"),
-    })
-}
-
-fn message_text(message: &AgentMessageInfo) -> Option<String> {
-    if !message.text.is_empty() {
-        return Some(message.text.clone());
-    }
-    let text = message
-        .parts
-        .iter()
-        .filter_map(message_part_text)
-        .collect::<Vec<_>>()
-        .join("");
-    if text.is_empty() { None } else { Some(text) }
-}
-
-fn message_part_text(part: &AgentMessagePartInfo) -> Option<String> {
-    if !part.text.trim().is_empty() {
-        return Some(part.text.clone());
-    }
-    if let Some(json) = part.json.as_ref() {
-        return compact_json(json).ok();
-    }
-    if let Some(tool_call) = part.tool_call.as_ref() {
-        return compact_json(tool_call)
-            .ok()
-            .map(|value| format!("tool call {value}"));
-    }
-    if let Some(tool_result) = part.tool_result.as_ref() {
-        return compact_json(tool_result)
-            .ok()
-            .map(|value| format!("tool result {value}"));
-    }
-    if let Some(image_ref) = part.image_ref.as_ref() {
-        return compact_json(image_ref)
-            .ok()
-            .map(|value| format!("image {value}"));
-    }
-    None
-}
-
-fn display_text(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    if display.text.is_empty() {
-        None
-    } else {
-        Some(&display.text)
-    }
-}
-
-fn rendered_display_text(display: &AgentTurnDisplayInfo) -> Option<String> {
-    let text = display_text(display)?;
-    Some(display_markdown::plain_text_for_format(
-        text,
-        display_format(display),
-        display_language(display),
-    ))
-}
-
-fn display_label(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    non_empty_str(&display.label)
-}
-
-fn display_ref(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    non_empty_str(&display.display_ref)
-}
-
-fn display_action(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    non_empty_str(&display.action)
-}
-
-fn display_format(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    non_empty_str(&display.format)
-}
-
-fn display_language(display: &AgentTurnDisplayInfo) -> Option<&str> {
-    non_empty_str(&display.language)
-}
-
-fn display_tool_label(_event: &AgentTurnEventInfo, display: &AgentTurnDisplayInfo) -> String {
-    display_label(display)
-        .or_else(|| display_ref(display))
-        .map(ToString::to_string)
-        .unwrap_or_else(|| "tool".to_string())
-}
-
-fn display_tool_ref(_event: &AgentTurnEventInfo, display: &AgentTurnDisplayInfo) -> Option<String> {
-    display_ref(display).map(ToString::to_string)
-}
-
-fn display_tool_input<'a>(
-    _event: &'a AgentTurnEventInfo,
-    display: &'a AgentTurnDisplayInfo,
-) -> Option<&'a Value> {
-    display.input.as_ref()
-}
-
-fn display_tool_output<'a>(
-    _event: &'a AgentTurnEventInfo,
-    display: &'a AgentTurnDisplayInfo,
-) -> Option<&'a Value> {
-    display.output.as_ref()
-}
-
-fn display_tool_error<'a>(
-    _event: &'a AgentTurnEventInfo,
-    display: &'a AgentTurnDisplayInfo,
-) -> Option<&'a Value> {
-    display.error.as_ref()
-}
-
-fn display_status(_event: &AgentTurnEventInfo, display: &AgentTurnDisplayInfo) -> Option<String> {
-    display_text(display).map(ToString::to_string)
-}
-
-fn display_value_text(value: &Value) -> Result<String> {
-    match value {
-        Value::String(text) => Ok(text.clone()),
-        _ => compact_json(value),
-    }
-}
-
-fn non_empty_str(value: &str) -> Option<&str> {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed)
-    }
-}
-
-fn compact_json(value: &Value) -> Result<String> {
-    serde_json::to_string(value).context("failed to encode compact JSON")
-}
-
-fn pretty_json(value: &Value) -> Result<String> {
-    serde_json::to_string_pretty(value).context("failed to encode pretty JSON")
-}
-
-fn build_session_create_body(args: &AgentSessionCreateArgs) -> Result<Value> {
-    let mut body = match args.input.as_deref() {
-        Some(path) => params::load_input_file(path)?,
-        None => Map::new(),
-    };
-
-    if let Some(provider) = args.provider.as_deref() {
-        body.insert("provider".to_string(), Value::String(provider.to_string()));
-    }
-    if let Some(model) = args.model.as_deref() {
-        body.insert("model".to_string(), Value::String(model.to_string()));
-    }
-    if let Some(client_ref) = args.client_ref.as_deref() {
-        body.insert(
-            "clientRef".to_string(),
-            Value::String(client_ref.to_string()),
-        );
-    }
-    if let Some(idempotency_key) = args.idempotency_key.as_deref() {
-        body.insert(
-            "idempotencyKey".to_string(),
-            Value::String(idempotency_key.to_string()),
-        );
-    }
-
-    Ok(Value::Object(body))
-}
-
-fn build_session_update_body(args: &AgentSessionUpdateArgs) -> Result<Value> {
-    let mut body = match args.input.as_deref() {
-        Some(path) => params::load_input_file(path)?,
-        None => Map::new(),
-    };
-
-    body.remove("provider");
-    body.remove("model");
-
-    if let Some(client_ref) = args.client_ref.as_deref() {
-        body.insert(
-            "clientRef".to_string(),
-            Value::String(client_ref.to_string()),
-        );
-    }
-    if let Some(state) = args.state.as_deref() {
-        body.insert("state".to_string(), Value::String(state.to_string()));
-    }
-
-    Ok(Value::Object(body))
-}
-
-fn build_turn_create_body(args: &AgentTurnCreateArgs) -> Result<Value> {
-    let mut body = match args.input.as_deref() {
-        Some(path) => params::load_input_file(path)?,
-        None => Map::new(),
-    };
-
-    body.remove("provider");
-    body.remove("clientRef");
-    body.remove("state");
-
-    if let Some(model) = args.model.as_deref() {
-        body.insert("model".to_string(), Value::String(model.to_string()));
-    }
-    if let Some(idempotency_key) = args.idempotency_key.as_deref() {
-        body.insert(
-            "idempotencyKey".to_string(),
-            Value::String(idempotency_key.to_string()),
-        );
-    }
-
-    let messages = build_messages(&args.system, &args.messages);
-    if !messages.is_empty() {
-        body.insert("messages".to_string(), Value::Array(messages));
-    }
-    if !args.tools.is_empty() {
-        body.insert(
-            "toolRefs".to_string(),
-            Value::Array(args.tools.iter().map(agent_tool_ref_value).collect()),
-        );
-    }
-
-    validate_turn_create_body(&body)?;
-    Ok(Value::Object(body))
-}
-
-fn validate_turn_create_body(body: &Map<String, Value>) -> Result<()> {
-    let has_messages = body
-        .get("messages")
-        .and_then(Value::as_array)
-        .is_some_and(|messages| !messages.is_empty());
-    if !has_messages {
-        bail!(
-            "agent turns create requires at least one message; pass --message, --system, or --input with a non-empty messages array"
-        );
-    }
-    Ok(())
-}
-
-fn build_messages(system: &[String], messages: &[String]) -> Vec<Value> {
-    let mut out = Vec::with_capacity(system.len() + messages.len());
-    for text in system {
-        out.push(json!({ "role": "system", "text": text }));
-    }
-    for text in messages {
-        out.push(json!({ "role": "user", "text": text }));
-    }
-    out
-}
-
-fn agent_tool_ref_value(tool: &AgentToolArg) -> Value {
-    json!({
-        "app": tool.app,
-        "operation": tool.operation,
-    })
-}
-
-fn sessions_path(
-    provider: Option<&str>,
-    state: Option<&str>,
-    summary_limit: Option<usize>,
-) -> String {
-    let mut serializer = url::form_urlencoded::Serializer::new(String::new());
-    if let Some(provider) = provider {
-        serializer.append_pair("provider", provider);
-    }
-    if let Some(state) = state {
-        serializer.append_pair("state", state);
-    }
-    if let Some(limit) = summary_limit {
-        serializer.append_pair("view", "summary");
-        serializer.append_pair("limit", &limit.to_string());
-    }
-    let query = serializer.finish();
-    if query.is_empty() {
-        SESSIONS_PATH.to_string()
-    } else {
-        format!("{SESSIONS_PATH}?{query}")
-    }
-}
-
-fn session_turns_path(session_id: &str, status: Option<&str>) -> String {
-    let mut serializer = url::form_urlencoded::Serializer::new(String::new());
-    if let Some(status) = status {
-        serializer.append_pair("status", status);
-    }
-    let query = serializer.finish();
-    let path = format!("{SESSIONS_PATH}/{session_id}/turns");
-    if query.is_empty() {
-        path
-    } else {
-        format!("{path}?{query}")
-    }
-}
-
-fn turn_events_path(
-    id: &str,
-    stream: bool,
-    after: Option<u64>,
-    limit: Option<u32>,
-    until: Option<&str>,
-) -> String {
-    let mut serializer = url::form_urlencoded::Serializer::new(String::new());
-    if let Some(after) = after {
-        serializer.append_pair("after", &after.to_string());
-    }
-    if let Some(limit) = limit {
-        serializer.append_pair("limit", &limit.to_string());
-    }
-    if let Some(until) = until {
-        serializer.append_pair("until", until);
-    }
-    let suffix = if stream { "/events/stream" } else { "/events" };
-    let query = serializer.finish();
-    let path = format!("{TURNS_PATH}/{id}{suffix}");
-    if query.is_empty() {
-        path
-    } else {
-        format!("{path}?{query}")
-    }
-}
-
-fn print_session(value: &Value, format: Format) {
-    match format {
-        Format::Json => output::print_json(value),
-        Format::Table => {
-            let rows = vec![session_row(value)];
-            output::print_table(&session_headers(), &rows);
-        }
-    }
-}
-
-fn print_sessions(value: &Value, format: Format) {
-    match format {
-        Format::Json => output::print_json(value),
-        Format::Table => {
-            let items = value.as_array().cloned().unwrap_or_default();
-            let rows: Vec<Vec<String>> = items.iter().map(session_row).collect();
-            output::print_table(&session_headers(), &rows);
-        }
-    }
-}
-
-fn print_turn(value: &Value, format: Format) {
-    match format {
-        Format::Json => output::print_json(value),
-        Format::Table => {
-            let rows = vec![turn_row(value)];
-            output::print_table(&turn_headers(), &rows);
-        }
-    }
-}
-
-fn print_turns(value: &Value, format: Format) {
-    match format {
-        Format::Json => output::print_json(value),
-        Format::Table => {
-            let items = value.as_array().cloned().unwrap_or_default();
-            let rows: Vec<Vec<String>> = items.iter().map(turn_row).collect();
-            output::print_table(&turn_headers(), &rows);
-        }
-    }
-}
-
-fn print_turn_events(value: &Value, format: Format) {
-    match format {
-        Format::Json => output::print_json(value),
-        Format::Table => {
-            let items = value.as_array().cloned().unwrap_or_default();
-            let rows: Vec<Vec<String>> = items.iter().map(turn_event_row).collect();
-            output::print_table(&event_headers(), &rows);
-        }
-    }
-}
-
-fn session_headers() -> [&'static str; 6] {
-    ["ID", "Provider", "Model", "State", "Client Ref", "Updated"]
-}
-
-fn turn_headers() -> [&'static str; 6] {
-    ["ID", "Session", "Provider", "Model", "Status", "Created"]
-}
-
-fn event_headers() -> [&'static str; 7] {
-    [
-        "Seq",
-        "Type",
-        "Source",
-        "Visibility",
-        "Turn",
-        "Created",
-        "Display",
-    ]
-}
-
-fn session_row(value: &Value) -> Vec<String> {
-    vec![
-        value["id"].as_str().unwrap_or("-").to_string(),
-        value["provider"].as_str().unwrap_or("-").to_string(),
-        value["model"].as_str().unwrap_or("-").to_string(),
-        value["state"].as_str().unwrap_or("-").to_string(),
-        value["clientRef"].as_str().unwrap_or("-").to_string(),
-        value["updatedAt"].as_str().unwrap_or("-").to_string(),
-    ]
-}
-
-fn turn_row(value: &Value) -> Vec<String> {
-    vec![
-        value["id"].as_str().unwrap_or("-").to_string(),
-        value["sessionId"].as_str().unwrap_or("-").to_string(),
-        value["provider"].as_str().unwrap_or("-").to_string(),
-        value["model"].as_str().unwrap_or("-").to_string(),
-        value["status"].as_str().unwrap_or("-").to_string(),
-        value["createdAt"].as_str().unwrap_or("-").to_string(),
-    ]
-}
-
-fn turn_event_row(value: &Value) -> Vec<String> {
-    let display = turn_event_display_summary(value)
-        .unwrap_or_else(|| fallback_turn_event_data_summary(value));
-    vec![
-        value["seq"]
-            .as_i64()
-            .map(|seq| seq.to_string())
-            .unwrap_or_else(|| "-".to_string()),
-        value["type"].as_str().unwrap_or("-").to_string(),
-        value["source"].as_str().unwrap_or("-").to_string(),
-        value["visibility"].as_str().unwrap_or("-").to_string(),
-        value["turnId"].as_str().unwrap_or("-").to_string(),
-        value["createdAt"].as_str().unwrap_or("-").to_string(),
-        display,
-    ]
-}
-
-fn fallback_turn_event_data_summary(value: &Value) -> String {
-    if let Ok(event) = serde_json::from_value::<AgentTurnEventInfo>(value.clone()) {
-        if let Some(summary) = turn_event_data_summary(&event) {
-            return summary;
-        }
-        if event.visibility == "private" {
-            return String::new();
-        }
-    } else if value["visibility"].as_str() == Some("private") {
-        return String::new();
-    }
-    serde_json::to_string(&value["data"]).unwrap_or_else(|_| "-".to_string())
-}
-
-fn turn_event_data_summary(event: &AgentTurnEventInfo) -> Option<String> {
-    match event.event_type.as_str() {
-        "agent.message.delta" | "assistant.delta" => {
-            string_any_field(&event.data, &["text", "delta", "content"])
-                .map(|text| format!("assistant delta: {text}"))
-        }
-        "assistant.completed" => {
-            string_field(&event.data, "text").map(|text| format!("assistant completed: {text}"))
-        }
-        "turn.started" => string_any_field(&event.data, &["status", "state"])
-            .map(|status| format!("turn started: {status}"))
-            .or_else(|| Some("turn started".to_string())),
-        "turn.completed" => string_any_field(&event.data, &["status", "state"])
-            .map(|status| format!("turn completed: {status}"))
-            .or_else(|| Some("turn completed".to_string())),
-        "turn.failed" => string_field(&event.data, "error")
-            .map(|error| format!("turn failed: {error}"))
-            .or_else(|| Some("turn failed".to_string())),
-        "turn.canceled" => string_field(&event.data, "reason")
-            .map(|reason| format!("turn canceled: {reason}"))
-            .or_else(|| Some("turn canceled".to_string())),
-        "tool.started" => Some(tool_data_summary(event, "started")?),
-        "tool.completed" => Some(tool_data_summary(event, "completed")?),
-        "tool.failed" => Some(tool_data_summary(event, "failed")?),
-        "interaction.requested" => {
-            let id = string_any_field(&event.data, &["interaction_id", "interactionId"])
-                .unwrap_or_else(|| "interaction".to_string());
-            Some(format!("interaction requested ({id})"))
-        }
-        "interaction.resolved" => {
-            let id = string_any_field(&event.data, &["interaction_id", "interactionId"])
-                .unwrap_or_else(|| "interaction".to_string());
-            Some(format!("interaction resolved ({id})"))
-        }
-        _ => None,
-    }
-}
-
-fn tool_data_summary(event: &AgentTurnEventInfo, phase: &str) -> Option<String> {
-    let tool = string_any_field(
-        &event.data,
-        &[
-            "tool_name",
-            "toolName",
-            "name",
-            "operation",
-            "tool_id",
-            "toolId",
-        ],
-    )
-    .unwrap_or_else(|| "tool".to_string());
-    let mut summary = format!("{tool} {phase}");
-    if let Some(status) = string_any_field(&event.data, &["status", "state"]).or_else(|| {
-        number_any_field(&event.data, &["status", "statusCode"]).map(|status| status.to_string())
-    }) {
-        summary.push_str(&format!(" ({status})"));
-    }
-    if let Some(error) = string_field(&event.data, "error") {
-        summary.push_str(&format!(": {error}"));
-    } else if let Some(output) = value_any_field(&event.data, &["output", "result", "body"]) {
-        summary.push(' ');
-        summary.push_str(&compact_json(output).ok()?);
-    } else if let Some(input) = value_any_field(&event.data, &["arguments", "input", "request"]) {
-        summary.push(' ');
-        summary.push_str(&compact_json(input).ok()?);
-    }
-    Some(summary)
-}
-
-fn turn_event_display_summary(value: &Value) -> Option<String> {
-    let event: AgentTurnEventInfo = serde_json::from_value(value.clone()).ok()?;
-    let display = turn_event_display(&event)?;
-    match display.kind.trim() {
-        "text" => display_text(display).map(|text| match display.phase.trim() {
-            "delta" => format!("assistant delta: {text}"),
-            "completed" => format!("assistant completed: {text}"),
-            phase if !phase.is_empty() => format!("assistant {phase}: {text}"),
-            _ => format!("assistant: {text}"),
-        }),
-        "reasoning" => display_text(display).map(|text| format!("reasoning: {text}")),
-        "tool" => Some(tool_display_summary(&event, display)?),
-        "interaction" => {
-            let label = display_label(display).unwrap_or("interaction");
-            let reference = display_ref(display).unwrap_or(label);
-            match display.phase.trim() {
-                "requested" => Some(format!("{label} requested ({reference})")),
-                "resolved" => Some(format!("{label} resolved ({reference})")),
-                phase if !phase.is_empty() => Some(format!("{label} {phase} ({reference})")),
-                _ => Some(label.to_string()),
-            }
-        }
-        "status" => {
-            let label = display_label(display).unwrap_or("turn");
-            match (display.phase.trim(), display_text(display)) {
-                ("started", Some(text)) => Some(format!("{label} started: {text}")),
-                ("started", None) => Some(format!("{label} started")),
-                ("completed", Some(text)) => Some(format!("{label} completed: {text}")),
-                ("completed", None) => Some(format!("{label} completed")),
-                ("canceled", Some(text)) => Some(format!("{label} canceled: {text}")),
-                ("canceled", None) => Some(format!("{label} canceled")),
-                ("progress", Some(text)) => Some(format!("{label}: {text}")),
-                (phase, Some(text)) if !phase.is_empty() => {
-                    Some(format!("{label} {phase}: {text}"))
-                }
-                _ => None,
-            }
-        }
-        "error" => {
-            let text = display_text(display).map(ToString::to_string).or_else(|| {
-                display
-                    .error
-                    .as_ref()
-                    .and_then(|value| display_value_text(value).ok())
-            })?;
-            let label = display_label(display).unwrap_or("error");
-            match display.phase.trim() {
-                "failed" if label == "turn" => Some(format!("turn failed: {text}")),
-                "failed" => Some(format!("{label} failed: {text}")),
-                _ => Some(format!("{label}: {text}")),
-            }
-        }
-        _ => None,
-    }
-}
-
-fn tool_display_summary(
-    event: &AgentTurnEventInfo,
-    display: &AgentTurnDisplayInfo,
-) -> Option<String> {
-    let tool = display_tool_label(event, display);
-    let mut summary = match display.phase.trim() {
-        "started" => display_action(display)
-            .map(|action| format!("{action} {tool}"))
-            .unwrap_or_else(|| format!("{tool} started")),
-        "completed" => display_action(display)
-            .map(|action| format!("{action} {tool}"))
-            .unwrap_or_else(|| format!("{tool} completed")),
-        "failed" => display_action(display)
-            .map(|action| format!("{action} {tool}"))
-            .unwrap_or_else(|| format!("{tool} failed")),
-        "progress" => match (display_action(display), display_text(display)) {
-            (Some(action), Some(text)) => format!("{action} {tool}: {text}"),
-            (Some(action), None) => format!("{action} {tool}"),
-            (None, Some(text)) => format!("{tool} {text}"),
-            (None, None) => format!("{tool} progress"),
-        },
-        phase if !phase.is_empty() => format!("{tool} {phase}"),
-        _ => tool,
-    };
-    if matches!(display.phase.trim(), "completed" | "failed")
-        && let Some(status) = display_status(event, display)
-    {
-        summary.push_str(&format!(" ({status})"));
-    }
-    if let Some(error) = display_tool_error(event, display) {
-        summary.push_str(&format!(": {}", display_value_text(error).ok()?));
-    } else if let Some(output) = display_tool_output(event, display) {
-        summary.push(' ');
-        summary.push_str(&compact_json(output).ok()?);
-    } else if let Some(input) = display_tool_input(event, display) {
-        summary.push(' ');
-        summary.push_str(&compact_json(input).ok()?);
-    }
-    Some(summary)
-}
