+++ gestalt/cli/src/commands/agents/fields/fixtures/assistant_delta.json
@@ -1,0 +1,3 @@
+{
+  "text": "hello world"
+}
