+++ gestalt/cli/src/commands/agents/fields/fixtures/interaction_requested.json
@@ -1,0 +1,3 @@
+{
+  "interactionId": "abc123"
+}
