+++ gestalt/cli/src/commands/agents/fields/fixtures/tool_started.json
@@ -1,0 +1,4 @@
+{
+  "tool_name": "grep",
+  "input": {"pattern": "foo"}
+}
