+++ gestalt/cli/src/commands/agents/mod.rs
@@ -1,0 +1,23 @@
+mod display_markdown;
+mod driver;
+mod fields;
+mod format;
+mod harness;
+mod render;
+mod requests;
+mod shell;
+mod stream;
+mod tui;
+mod types;
+
+pub use harness::{doctor_local, launch_local};
+pub use requests::{
+    cancel_turn, create_session, create_turn, get_session, get_turn, list_sessions,
+    list_turn_events, list_turns, stream_turn_events, transcript_turn, update_session,
+};
+pub use shell::{resume_interactive, run_interactive};
+
+pub(crate) use fields::compact_json;
+pub(crate) use requests::{cancel_turn_silent, create_turn_info, INTERRUPT_CANCEL_REASON};
+pub(crate) use shell::{agent_model_lines, agent_session_lines, agent_tui_help_lines};
+pub(crate) use types::{AgentInteractionInfo, AgentShell};
