+++ gestalt/cli/src/commands/agents/render/mod.rs
@@ -1,0 +1,1 @@
+pub mod stdout;
