+++ gestalt/cli/src/commands/agents/tui/worker.rs
@@ -1,18 +1,20 @@
 use std::sync::mpsc::{Receiver, Sender, TryRecvError};
 use std::thread;
-use std::time::Duration;
 
-use anyhow::{Context, Result, anyhow};
+use anyhow::{Context, Result};
 use serde_json::{Map, Value};
 
 use crate::api::ApiClient;
 use crate::cli::AgentTurnCreateArgs;
-
-use super::super::{
-    AgentInteractionInfo, AgentTurnEventInfo, AgentTurnInfo, INTERRUPT_CANCEL_REASON,
-    cancel_turn_silent, create_turn_info, get_turn_info, list_interactions_info,
-    resolve_interaction_info, stream_turn_event_frames,
+use crate::commands::agents::driver::{
+    TurnDriverSink, TurnLoopContext, TurnLoopOutcome, resolve_pending_interactions,
+    run_turn_status_loop,
 };
+use crate::commands::agents::requests::{
+    INTERRUPT_CANCEL_REASON, cancel_turn_silent, create_turn_info, resolve_interaction_info,
+};
+use crate::commands::agents::stream::stream_turn_event_frames;
+use crate::commands::agents::types::{AgentInteractionInfo, AgentTurnEventInfo, AgentTurnInfo};
 
 pub(super) struct TurnWorker {
     pub(super) command_tx: Sender<WorkerCommand>,
@@ -63,66 +65,58 @@ fn run_turn_worker(
     event_tx
         .send(WorkerEvent::TurnCreated(turn))
         .context("terminal UI closed before turn started")?;
-    if consume_cancel(client, &turn_id, command_rx) {
-        return Ok(());
-    }
 
-    let mut after_seq = 0;
-    loop {
-        stream_turn_event_frames(client, &turn_id, after_seq, |event| {
+    let ctx = TurnLoopContext {
+        client,
+        turn_id: &turn_id,
+    };
+    let mut sink = WorkerTurnDriver {
+        after_seq: 0,
+        event_tx,
+        command_rx,
+    };
+    run_turn_status_loop(&ctx, &mut sink)
+}
+
+struct WorkerTurnDriver<'a> {
+    after_seq: u64,
+    event_tx: &'a Sender<WorkerEvent>,
+    command_rx: &'a Receiver<WorkerCommand>,
+}
+
+impl TurnDriverSink for WorkerTurnDriver<'_> {
+    fn stream_events(&mut self, ctx: &TurnLoopContext<'_>) -> Result<()> {
+        stream_turn_event_frames(ctx.client, ctx.turn_id, self.after_seq, |event| {
             if event.seq > 0 {
-                after_seq = after_seq.max(event.seq as u64);
+                self.after_seq = self.after_seq.max(event.seq as u64);
             }
-            event_tx
+            self.event_tx
                 .send(WorkerEvent::TurnEvent(Box::new(event)))
                 .context("terminal UI closed while streaming events")
-        })?;
+        })
+    }
 
-        if consume_cancel(client, &turn_id, command_rx) {
-            return Ok(());
-        }
+    fn on_turn_snapshot(&mut self, turn: &AgentTurnInfo) -> Result<()> {
+        self.event_tx
+            .send(WorkerEvent::TurnSnapshot(turn.clone()))
+            .context("terminal UI closed before turn snapshot")
+    }
 
-        let latest = get_turn_info(client, &turn_id)?;
-        event_tx
-            .send(WorkerEvent::TurnSnapshot(latest.clone()))
-            .context("terminal UI closed before turn snapshot")?;
+    fn handle_waiting_for_input(
+        &mut self,
+        ctx: &TurnLoopContext<'_>,
+        turn: &AgentTurnInfo,
+    ) -> Result<TurnLoopOutcome> {
+        resolve_pending_interactions(ctx, turn, |interaction| {
+            self.event_tx
+                .send(WorkerEvent::WaitingForInput(interaction.clone()))
+                .context("terminal UI closed before interaction prompt")?;
+            wait_for_interaction_resolution(ctx, &interaction, self.command_rx, self.event_tx)
+        })
+    }
 
-        match latest.status.as_str() {
-            "waiting_for_input" => {
-                let interactions = list_interactions_info(client, &turn_id)?;
-                let pending: Vec<_> = interactions
-                    .into_iter()
-                    .filter(|interaction| interaction.state == "pending")
-                    .collect();
-                if pending.is_empty() {
-                    return Err(anyhow!(
-                        "agent turn {} is waiting for input without a pending interaction",
-                        latest.id
-                    ));
-                }
-                for interaction in pending {
-                    event_tx
-                        .send(WorkerEvent::WaitingForInput(interaction.clone()))
-                        .context("terminal UI closed before interaction prompt")?;
-                    wait_for_interaction_resolution(
-                        client,
-                        &turn_id,
-                        interaction,
-                        command_rx,
-                        event_tx,
-                    )?;
-                }
-            }
-            "pending" | "running" => thread::sleep(Duration::from_millis(250)),
-            "succeeded" | "failed" | "canceled" => return Ok(()),
-            other => {
-                return Err(anyhow!(
-                    "agent turn {} has unsupported status {}",
-                    latest.id,
-                    other
-                ));
-            }
-        }
+    fn poll_cancel(&mut self, ctx: &TurnLoopContext<'_>) -> Result<bool> {
+        Ok(consume_cancel(ctx.client, ctx.turn_id, self.command_rx))
     }
 }
 
@@ -138,31 +132,34 @@ fn consume_cancel(client: &ApiClient, turn_id: &str, command_rx: &Receiver<Worke
 }
 
 fn wait_for_interaction_resolution(
-    client: &ApiClient,
-    turn_id: &str,
-    interaction: AgentInteractionInfo,
+    ctx: &TurnLoopContext<'_>,
+    interaction: &AgentInteractionInfo,
     command_rx: &Receiver<WorkerCommand>,
     event_tx: &Sender<WorkerEvent>,
-) -> Result<()> {
+) -> Result<TurnLoopOutcome> {
     loop {
         match command_rx
             .recv()
             .context("terminal UI closed before resolving interaction")?
         {
             WorkerCommand::Cancel => {
-                let _ = cancel_turn_silent(client, turn_id, INTERRUPT_CANCEL_REASON);
-                return Ok(());
+                let _ = cancel_turn_silent(ctx.client, ctx.turn_id, INTERRUPT_CANCEL_REASON);
+                return Ok(TurnLoopOutcome::Cancelled);
             }
             WorkerCommand::Resolve {
                 interaction_id,
                 resolution,
             } if interaction_id == interaction.id => {
-                let resolved =
-                    resolve_interaction_info(client, turn_id, &interaction.id, resolution)?;
+                let resolved = resolve_interaction_info(
+                    ctx.client,
+                    ctx.turn_id,
+                    &interaction.id,
+                    resolution,
+                )?;
                 event_tx
                     .send(WorkerEvent::InteractionResolved(resolved))
                     .context("terminal UI closed after resolving interaction")?;
-                return Ok(());
+                return Ok(TurnLoopOutcome::Continue);
             }
             WorkerCommand::Resolve { .. } => {}
         }
