+++ gestalt/cli/src/commands/authorization.rs
@@ -18,6 +18,7 @@ use crate::cli::{
 };
 use crate::commands::apps;
 use crate::output::{self, Format};
+use crate::query;
 
 const SUBJECT_PREFIX: &str = "service_account:";
 
@@ -893,29 +894,15 @@ fn read_input(path: &str) -> Result<String> {
 }
 
 fn append_query(path: &str, params: &[(String, String)]) -> Result<String> {
-    if params.is_empty() {
-        return Ok(path.to_string());
-    }
-    Ok(format!(
-        "{}?{}",
-        path,
-        serde_urlencoded::to_string(params).context("failed to encode query")?
-    ))
+    query::append_query(path, params)
 }
 
 fn page_params(page_size: Option<u32>, page_token: Option<&str>) -> Vec<(String, String)> {
-    let mut params = Vec::new();
-    if let Some(page_size) = page_size {
-        params.push(("pageSize".to_string(), page_size.to_string()));
-    }
-    push_param(&mut params, "pageToken", page_token);
-    params
+    query::page_params(page_size, page_token)
 }
 
 fn push_param(params: &mut Vec<(String, String)>, name: &str, value: Option<&str>) {
-    if let Some(value) = value.map(str::trim).filter(|value| !value.is_empty()) {
-        params.push((name.to_string(), value.to_string()));
-    }
+    query::push_opt_param(params, name, value);
 }
 
 fn object() -> Map<String, Value> {
